tests: split monolithic test.mjs into focused modules
- Add helpers.mjs with fixtures() path helper and mock setup functions
- Split into dependency-tree, formats, to-list, resolvers, node-modules,
  and config test files
- Add coverage for Sass nonExistent partials, _getDependencies parse errors,
  flat/scoped node_modules, requirejs nonExistent aliases, ES6 lazy/dynamic
  imports, nodeModulesConfig.entry, tsconfig/tsx/allowJs integration
- Flatten nested describes, inline single-use beforeEach() mocks, lift
  helpers to module scope, fix shadowed variables
