+++ package.json
@@ -16,7 +16,7 @@
   "scripts": {
     "lint": "xo",
     "fix": "xo --fix",
-    "mocha": "mocha",
+    "mocha": "mocha \"test/*.test.mjs\"",
     "test": "npm run lint && npm run mocha",
     "test:ci": "c8 npm run mocha"
   },
@@ -103,6 +103,14 @@
       "unicorn/prefer-module": "off",
       "unicorn/prefer-top-level-await": "off",
       "unicorn/prevent-abbreviations": "off"
-    }
+    },
+    "overrides": [
+      {
+        "files": "test/**/*.{c,m}js",
+        "envs": [
+          "mocha"
+        ]
+      }
+    ]
   }
 }
